Move curly brace to newline to conform to destination interface coding styles.
There are differences in coding styles between the source module, where the implementations originated, and the destination module, where the default methods were created. This commit is to conform to to new coding style. This is a manual step as we do not currently analyze the difference in coding styles between projects/modules.

Signed-off-by: Raffi Khatchadourian <rkhatchadourian@citytech.cuny.edu>