Clean up BND instructions
* Remove jcni version (it's been removed in 2017 from the code base).
* Remove api version (it's never been used since moving the code to
eclipse).
* Set a fixed Bundle-Developer in generated manifests to avoid
bnd-maven-plugin complaining about all the developers in the parent
pom.xml not having a mandatory ID.
* Set the Bundle-Description header in BND. Otherwise the pom.xml
description is taken for that header, and unfortunately the pom.xml
description tag contains whitespace from the XML formatting and
indenting.
* Use the versionmask macro instead of the version macro. version is an
alias for versionmask that is for compatibility only and should not be
used.